+++ src/cmd/bots/rag/data_process/youtube_audio_transcribe_to_tidb.py
@@ -6,18 +6,14 @@ import os
 import time
 import subprocess
 
-
-from pytube import YouTube, cipher
-from pytube.innertube import _default_clients
-from pytube.exceptions import RegexMatchError
 from deep_translator import GoogleTranslator
-
 from deepgram import DeepgramClient, PrerecordedOptions, FileSource
 import httpx
 from langchain_text_splitters import RecursiveCharacterTextSplitter
 from langchain_community.vectorstores import TiDBVectorStore
 from langchain_community.embeddings import JinaEmbeddings
 
+from .pytube_monkey_fix import YouTube
 from src.common.types import VIDEOS_DIR
 from src.common.logger import Logger
 from src.cmd.bots.rag.helper import get_tidb_url
@@ -59,60 +55,6 @@ def get_video_duration(video_file):
     duration = sed_process.communicate()[0].decode('utf-8').strip()
 
     return duration
-
-
-# issue: https://github.com/pytube/pytube/issues/1973
-_default_clients["ANDROID"]["context"]["client"]["clientVersion"] = "19.08.35"
-_default_clients["IOS"]["context"]["client"]["clientVersion"] = "19.08.35"
-_default_clients["ANDROID_EMBED"]["context"]["client"]["clientVersion"] = "19.08.35"
-_default_clients["IOS_EMBED"]["context"]["client"]["clientVersion"] = "19.08.35"
-_default_clients["IOS_MUSIC"]["context"]["client"]["clientVersion"] = "6.41"
-_default_clients["ANDROID_MUSIC"] = _default_clients["ANDROID_CREATOR"]
-
-
-def get_throttling_function_name(js: str) -> str:
-    """Extract the name of the function that computes the throttling parameter.
-
-    :param str js:
-        The contents of the base.js asset file.
-    :rtype: str
-    :returns:
-        The name of the function used to compute the throttling parameter.
-    """
-    function_patterns = [
-        r'a\.[a-zA-Z]\s*&&\s*\([a-z]\s*=\s*a\.get\("n"\)\)\s*&&.*?\|\|\s*([a-z]+)',
-        r'a\.[a-zA-Z]\s*&&\s*\([a-z]\s*=\s*a\.get\("n"\)\)\s*&&\s*'
-        r'\([a-z]\s*=\s*([a-zA-Z0-9$]+)(\[\d+\])?\([a-z]\)',
-        r'\([a-z]\s*=\s*([a-zA-Z0-9$]+)(\[\d+\])\([a-z]\)',
-
-    ]
-    # logging.debug('Finding throttling function name')
-    for pattern in function_patterns:
-        regex = re.compile(pattern)
-        function_match = regex.search(js)
-        if function_match:
-            logging.debug("finished regex search, matched: %s", pattern)
-            if len(function_match.groups()) == 1:
-                return function_match.group(1)
-            idx = function_match.group(2)
-            if idx:
-                idx = idx.strip("[]")
-                array = re.search(
-                    r'var {nfunc}\s*=\s*(\[.+?\]);'.format(
-                        nfunc=re.escape(function_match.group(1))),
-                    js
-                )
-                if array:
-                    array = array.group(1).strip("[]").split(",")
-                    array = [x.strip() for x in array]
-                    return array[int(idx)]
-
-    raise RegexMatchError(
-        caller="get_throttling_function_name", pattern="multiple"
-    )
-
-
-cipher.get_throttling_function_name = get_throttling_function_name
 
 
 def download_videos(link: str, download_path: str):
